0.9.0 Release

+++ bin/actions/uninstall_archive_action.js
@@ -6,9 +6,13 @@ const uninstallArchiveAction = (archive) => {
   return new Promise((resolve, reject) => {
     const uninstallPath = `${ApplicationConfiguration.get(`paths.archive`)}/${archive}`
 
-    FileSystem.removeDirectory(uninstallPath)
-    StatusLog.notify(`uninstalled`, archive)
-    resolve()
+    if (FileSystem.isDirectory(uninstallPath)) {
+      FileSystem.removeDirectory(uninstallPath)
+      StatusLog.notify(`uninstalled`, archive)
+      resolve()
+    } else {
+      reject(new Error(`${uninstallPath} does not exist`))
+    }
   })
 }
